[refactor/#14] Change package structure
