+++ src/main/java/com/api/meetudy/interest/entity/MemberInterest.java
@@ -1,4 +1,4 @@
-package com.api.meetudy.study.recommendation.entity;
+package com.api.meetudy.interest.entity;
 
 import com.api.meetudy.member.entity.Member;
 import jakarta.persistence.*;
